+++ ParkSpotTLV.Infrastructure/Entities/StreetSegment.cs
@@ -4,14 +4,14 @@ using System.ComponentModel.DataAnnotations;
 *****Paid Street Types******
 There are effectively two categories of paid parking (your PaidA and PaidB):
 
-PaidA (Standard Paid Parking – most zones) Zones 1,2,4,12,13
+PaidA (Standard Paid Parking – most zones) Zones 1,2,4,12,13 - 08:00–17:00 (Mon–Thu), 08:00–17:00 (Fri) 7 ₪ / 4.90 ₪ per hour
     Payment required Sunday–Thursday 08:00–19:00.
     Fridays and holiday eves 08:00–17:00.
     Price: 7 ₪ per hour if no permit at all for Tel-Aviv. If have SOME zone permit -> 4.90 ₪
     Residents of the zone do not pay.
     Applies to zones 1–7, except where special rules exist.
 
-PaidB (Extended Paid Parking – Central areas) - Zones 6,7,9,10
+PaidB (Extended Paid Parking – Central areas) - Zones 6,7,9,10 - 08:00–21:00 (Mon–Thu), 08:00–17:00 (Fri) 12.40 ₪ / 8.68 ₪ per hour
     Payment required Sunday–Thursday 08:00–21:00.
     Fridays and holiday eves 08:00–17:00.
     Price: 12.40 ₪ per hour if no permit at all for Tel-Aviv. If have SOME zone permit -> 8.68 ₪
@@ -22,29 +22,24 @@ PaidB (Extended Paid Parking – Central areas) - Zones 6,7,9,10
 
 namespace ParkSpotTLV.Infrastructure.Entities {
     public enum ParkingType {
-        Unknown = 0,
-        Free = 1,       // Always free
-        PaidA = 2,      // 08:00–17:00 (Mon–Thu), 08:00–17:00 (Fri) 7 ₪ / 4.90 ₪ per hour
-        PaidB = 3,      // 08:00–23:00 (Mon–Thu), 08:00–17:00 (Fri) 12.40 ₪ / 8.68 ₪ per hour
-        Limited = 4     // Time-restricted or other exceptions
-    }                
+        CantPark = 0,
+        Free = 1,
+        Paid = 2
+    }
 
-    // Which curb side the rule applies to
-    public enum SegmentSide { Both = 0, Left = 1, Right = 2 }
+    public enum SegmentSide { Both = 0, Left = 1, Right = 2 }   // Why is this important ?
 
     public class StreetSegment {
-        
+
         public Guid Id { get; set; } = Guid.NewGuid();
-        [MaxLength(128)] public string? Name { get; set; }
+        [Required] public string OSMId { get; set; } = "";
+        [MaxLength(128)] public string? NameEnglish { get; set; }
+        [MaxLength(128)] public string? NameHebrew { get; set; }
         [Required] public LineString Geom { get; set; } = default!;
         public Guid? ZoneId { get; set; }
         public Zone? Zone { get; set; }
-        public bool CarsOnly { get; set; } = false;
-        public ParkingType ParkingType { get; set; } = ParkingType.Unknown;
+        public ParkingType ParkingType { get; set; } = ParkingType.CantPark;
         public SegmentSide Side { get; set; } = SegmentSide.Both;
         public DateTimeOffset? LastUpdated { get; set; }
     }
-
 }
-
-